update file structure: clear separation of wrapper and msgpack_pack_format
