+++ src/wrapper/countminsketch_topk.rs
@@ -20,7 +20,6 @@ use crate::{DataInput, RegularPath, Vector2D};
 
 use serde::{Deserialize, Serialize};
 
-use crate::message_pack_format::dto::{CountMinSketchInnerWire, CountMinSketchWithHeapWire};
 use crate::message_pack_format::{Error as MsgPackError, MessagePackCodec};
 
 // ----- asap_sketchlib-backed CMSHeap helpers -----
@@ -308,50 +307,6 @@ impl CountMinSketchWithHeap {
             sketch.update(key, value);
         }
         sketch.serialize_msgpack().ok()
-    }
-}
-
-impl MessagePackCodec for CountMinSketchWithHeap {
-    fn to_msgpack(&self) -> Result<Vec<u8>, MsgPackError> {
-        let wire = CountMinSketchWithHeapWire {
-            sketch: CountMinSketchInnerWire {
-                sketch: self.sketch_matrix(),
-                rows: self.rows,
-                cols: self.cols,
-            },
-            topk_heap: self.topk_heap_items(),
-            heap_size: self.heap_size,
-        };
-        Ok(rmp_serde::to_vec(&wire)?)
-    }
-
-    fn from_msgpack(bytes: &[u8]) -> Result<Self, MsgPackError> {
-        let wire: CountMinSketchWithHeapWire = rmp_serde::from_slice(bytes)?;
-
-        let mut sorted_topk_heap = wire.topk_heap;
-        sorted_topk_heap.sort_by(|a, b| b.value.partial_cmp(&a.value).unwrap());
-
-        let wire_heap: Vec<WireHeapItem> = sorted_topk_heap
-            .iter()
-            .map(|h| WireHeapItem {
-                key: h.key.clone(),
-                value: h.value,
-            })
-            .collect();
-        let backend = sketchlib_cms_heap_from_matrix_and_heap(
-            wire.sketch.rows,
-            wire.sketch.cols,
-            wire.heap_size,
-            &wire.sketch.sketch,
-            &wire_heap,
-        );
-
-        Ok(Self {
-            rows: wire.sketch.rows,
-            cols: wire.sketch.cols,
-            heap_size: wire.heap_size,
-            backend,
-        })
     }
 }
