+++ src/wrapper/delta_set_aggregator.rs
@@ -8,19 +8,9 @@ use std::collections::HashSet;
 use crate::message_pack_format::{Error as MsgPackError, MessagePackCodec};
 
 /// Re-export of the wire DTO — canonical definition lives in
-/// [`crate::message_pack_format::dto::DeltaResult`]. Preserved here for
-/// backwards compatibility.
-pub use crate::message_pack_format::dto::DeltaResult;
-
-impl MessagePackCodec for DeltaResult {
-    fn to_msgpack(&self) -> Result<Vec<u8>, MsgPackError> {
-        Ok(rmp_serde::to_vec(self)?)
-    }
-
-    fn from_msgpack(bytes: &[u8]) -> Result<Self, MsgPackError> {
-        Ok(rmp_serde::from_slice(bytes)?)
-    }
-}
+/// [`crate::message_pack_format::delta_set_aggregator::DeltaResult`].
+/// Preserved here for backwards compatibility.
+pub use crate::message_pack_format::delta_set_aggregator::DeltaResult;
 
 /// Serialize a delta result to MessagePack. Thin shim over
 /// [`MessagePackCodec::to_msgpack`] preserved for backwards compatibility.
