+++ src/wrapper/hll.rs
@@ -297,16 +297,6 @@ impl HllSketch {
     }
 }
 
-impl MessagePackCodec for HllSketch {
-    fn to_msgpack(&self) -> Result<Vec<u8>, MsgPackError> {
-        Ok(rmp_serde::to_vec(self)?)
-    }
-
-    fn from_msgpack(bytes: &[u8]) -> Result<Self, MsgPackError> {
-        Ok(rmp_serde::from_slice(bytes)?)
-    }
-}
-
 #[cfg(test)]
 mod tests_wire_hll {
     use super::*;
